refactor block patterns into scripts package, move functionality into pattern_cutting package structure
