+++ scripts/TheClassicTailoredTrouserBlock.py
@@ -1,4 +1,4 @@
-from BodyMeasurements import BodyMeasurements, exampleBodyMeasurements
+from pattern_cutting.BodyMeasurements import BodyMeasurements, exampleBodyMeasurements
 from pattern_cutting.geometry.Group import Group
 from pattern_cutting.geometry.Shape import Shape
 from pattern_cutting.geometry.Vector import Vector, midpoint
