+++ scripts/curly-flare-pattern-block.py
