+++ src/pattern_cutting/BodyMeasurements.py
